refactor: move getIn to new CollectionUtil class
