+++ lib/src/map_extensions.dart
@@ -1,4 +1,4 @@
-import 'collections.dart' as coll;
+import 'collection_util.dart';
 import 'map_util.dart';
 
 /// Useful Map extensions
@@ -11,6 +11,6 @@ extension MapExtentions on Map {
   /// Takes a [path] `List` and returns the value at [path] in a nested collection.
   /// Traverses `Map` (by key), `List` (by index), and `Iterable` (by index) values.
   dynamic getIn(List<dynamic> path) {
-    return coll.getIn(path, this);
+    return CollectionUtil.getIn(path, this);
   }
 }
