+++ test/collections_test.dart
@@ -25,62 +25,4 @@ void main() {
     expect(pouch.isNotEmpty(''), isFalse);
     expect(pouch.isNotEmpty('1234'), isTrue);
   });
-
-  group('getIn()', () {
-    test('returns the value at path', () {
-      expect(pouch.getIn([1], [0, 1, 2]), equals(1));
-      {
-        var path = [1, 1];
-        var coll = [
-          0,
-          [4, 5, 6],
-          2
-        ];
-        expect(pouch.getIn(path, coll), equals(5));
-      }
-      {
-        var path = [1, 1, 'b'];
-        var coll = [
-          0,
-          [
-            4,
-            {'a': 'a', 'b': 'b'},
-            6
-          ],
-          2
-        ];
-        expect(pouch.getIn(path, coll), equals('b'));
-      }
-    });
-    test('errors if key is not found', () {
-      expect(
-        () => pouch.getIn([0], {1: 1}),
-        throwsA(isA<pouch.KeyNotFoundException>()),
-      );
-    });
-    test('errors if index is outside range', () {
-      expect(
-        () => pouch.getIn([2], [0, 1]),
-        throwsA(isA<RangeError>()),
-      );
-    });
-    test('errors if index is not an int', () {
-      expect(
-        () {
-          pouch.getIn(['a'], [0]);
-        },
-        throwsA(isA<TypeError>()),
-      );
-    });
-    test('returns null if collection is empty', () {
-      expect(pouch.getIn([1], []), isNull);
-      expect(pouch.getIn([1], {}), isNull);
-      expect(pouch.getIn([0, 0], [{}]), isNull);
-    });
-    test('returns null if `coll` is not a collection', () {
-      expect(pouch.getIn([1], 1), isNull);
-      expect(pouch.getIn([1], '1'), isNull);
-      expect(pouch.getIn([0, 0], [0]), isNull);
-    });
-  });
 }
